% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{A function for ...}
\usage{
analyze(db = NULL, exp, prop = "n", filterchk = NULL,
  basement = NULL, ifCluster = "both", includeORexclude = "1",
  esetName, ifLR = "0", path_to_write, up = 4000, down = 1000,
  window.size = 50, display = "all", debuging = FALSE)
}
\arguments{
\item{db}{db is the database name in the GenomicsPortals.}

\item{exp}{This is basically the experiment name.}

\item{prop}{is the property selected from the experiment.}

\item{filterchk}{filterchk is in the form of a string. It can have multiple property of the pData to filter
based on. Each pair of property:value should be separated by a comma ",". Pairs are saparated
by colon ":". An example ot filterchk is :
filterchk="property1:value1,property1:value2,property2:value1"}

\item{basement}{basement Levels of comparison, not relevant here!}

\item{ifCluster}{This parameter defines in which way the data should be clustered.}

\item{includeORexclude}{This argument basically is designed to filter the pData based on what 
filterchk is or selected property of eset when is set to "1", or the reverse selection of filterchk 
when set to "2". 
It also can have NULL or "n" values when there is no property selected, in this case 
the function will return the original ExpressionSet.}

\item{path_to_write}{This parameter specifies the path which user wants to save the results.}

\item{up, down}{Specifies the maximum number of up and down regulated genes.}

\item{window.size}{...}

\item{display}{Which resalts to show. It should be a vector of column indices or "all" for everything.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- analyze(exp="TCGA_KIRC_RNASeqV2", filterchk="n", prop="n", ifCluster="both", includeORexclude="2", 
		path_to_write= "/mnt/raid/tmp/", esetName="TCGA_BRCA_RNASeqV2_tcgaIlluminaHiSeq_Wed_Sep_13_12_30_57_2017", display=c(1,6,10))
}
\keyword{filter}
