% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFeatures.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{A function for ...}
\usage{
annotateFeatures(exp, flist = NULL, debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- annotateFeatures("EDS-1014")
}
\keyword{...}
