% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTregGRS.R
\name{batchTregGRS}
\alias{batchTregGRS}
\title{A function for ...}
\usage{
batchTregGRS(query.p, reference.p, query.gL = NULL,
  reference.gL = NULL, na.rm = TRUE, estimateNullDistr = TRUE,
  nullDistrQuantiles = c(0.9, 0.95, 0.99), nullDistrN = 100,
  tolerateWarnings = TRUE, pAdjust.method.query = NULL,
  pAdjust.method.reference = NULL, lambda = 0.005, rescale = T,
  plotRescaled = F, multicoreN = 3)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
