% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_datasets.R
\name{cluster_datasets}
\alias{cluster_datasets}
\title{A function for ...}
\usage{
cluster_datasets(sessionID, dataClust, nFactorLevels, ifCluster,
  currRowOrder, rowLable, rowsepindex, colsepindex, rsd, pal,
  prop_distance, isPropNone, isCentered, path_to_write, prop, factorLevels,
  allColors, nColors, sigDB = "ilincs_sigs", chost = "ilincs.org",
  debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
