% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenelistEnrichment.R
\name{computeGenelistEnrichment}
\alias{computeGenelistEnrichment}
\title{A function for ...}
\usage{
computeGenelistEnrichment(queryGenelist, libName, ncors = 10,
  debuging = FALSE, sigDB = "ilincs_sigs",
  chost = "gimm12.ketl.uc.edu", dhost = "gimm2.ketl.uc.edu",
  dport = 3306, org = "Hs")
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
