% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_with_pval.R
\name{cor_with_pval}
\alias{cor_with_pval}
\title{A function to create the concordance table between a signature and a library.}
\usage{
cor_with_pval(data1, data2, cortype = "pearson",
  alternative = "two.sided", topSigs = FALSE)
}
\arguments{
\item{data1}{A numeric vector as the input signature.}

\item{data2}{A numeric vector or a matrix in the same row orther as in data1.}

\item{cortype}{The correlation type like "pearson" or "spearman".}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- cor_with_pval(data1, data2)
}
\keyword{...}
