% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expr_for_gff_sample.R
\name{create_expr_for_gff_sample}
\alias{create_expr_for_gff_sample}
\title{A function for ...}
\usage{
create_expr_for_gff_sample(result, up, down, window.size = 100,
  missingWindow = 0)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
