% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jnlps.R
\name{create_jnlps}
\alias{create_jnlps}
\title{A function for ...}
\usage{
create_jnlps(sessionID, isCentered, path_to_write, sigDB = "ilincs_sigs",
  chost = "ilincs.org", debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
