% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEsetInfo.R
\name{createEsetInfo}
\alias{createEsetInfo}
\title{A function for ...}
\usage{
createEsetInfo(exp, glist, path_to_write = "/mnt/raid/tmp/", up = 4000,
  down = 1000, window.size = 50, debuging = FALSE)
}
\arguments{
\item{exp}{This is basically the experiment name.}

\item{path_to_write}{This parameter specifies the path which user wants to save the results.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- createEsetInfo(....)
}
\keyword{...}
