% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSignatureByDiffGEA.R
\name{createSignatureByDiffGEA}
\alias{createSignatureByDiffGEA}
\title{A function for ...}
\usage{
createSignatureByDiffGEA(exp, prop = NULL, filterchk = "n",
  basement = NULL, includeORexclude = 2, ifCluster = "Genes",
  up = 4000, down = 1000, window.size = 50, path_to_write, pORqVal,
  userUploadedProfilePath = NULL, previousID = NULL,
  reAnalyse = FALSE, glist = NULL, pValue = NULL,
  foldchange = NULL, MLinfo = NULL, write = FALSE,
  homoloGenes = TRUE, multigroup = FALSE, sigDB = "ilincs_sigs",
  chost = "ilincs.org", debuging = FALSE)
}
\arguments{
\item{basement}{A two value vector containing "baseline" and "treatment".}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
