% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSignature.R
\name{downloadSignature}
\alias{downloadSignature}
\title{A function to download Signatures.}
\usage{
downloadSignature(sigID, path_to_write = "/mnt/raid/tmp/",
  display = TRUE, geneNames = TRUE, pValue = NULL,
  foldchange = NULL, sigDB = "ilincs_sigs", species = "Hs",
  noOfGenes = "all", glist = NULL, write = TRUE, debuging = FALSE)
}
\arguments{
\item{sigID}{This is a string object. A signature ID or a list of signature IDs separated by commas.}

\item{path_to_write}{The path to save the signature/s as a ".csv" file. The default is set to a temp folder: "/mnt/raid/tmp/"}

\item{display}{If user wants to get the signatures as an R object (data.frame) otherwise it will only be saved.}

\item{geneNames}{If user wants to get gene names along with Entrez IDs in the downloaded signature(s).}

\item{pValue}{Cut off based on p-value.}

\item{foldchange}{Cut off based on fold change.}

\item{sigDB}{Which signature version user wants to use in the analysis.}

\item{noOfGenes}{Number of top genes in the signature based on their p-values.}

\item{glist}{Subset the signature based on a gene list.}

\item{write}{If need to save the tab delimited file.}

\item{debuging}{This is for debuging purposes only.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to download a list of Signatures from Genomics Portals database.
}
\examples{
## not run
res <- downloadSignature(sigID="ENC_1,ENC_10,ENC_15", path_to_write="my/path/to/save/", display=TRUE, debuging=FALSE)
## end not run
}
\keyword{Signature}
