% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugTargets.R
\name{drugTargets}
\alias{drugTargets}
\title{A function to extract gene targets for a drug.}
\usage{
drugTargets(lsmID, drugID = NULL, signatureID = NULL,
  sigDB = "ilincs_sigs", debuging = FALSE)
}
\arguments{
\item{lsmID}{A vector of LSM IDs.}

\item{drugID}{A vector of BRD IDs (perturbagenIDs).}

\item{signatureID}{A vector of chemical perturbage (CP) signature IDs.}

\item{sigDB}{Signature database to use.}

\item{debuging}{For debuging purposes only}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to extract gene targets related to specified drugID, lsmID or signatureID from ilincs portal.
}
\examples{
## not run
targets <- drugTargets(lsmID=c("LSM-45711", "LSM-5242", "LSM-45954"))
## end not run
}
\keyword{Signature,}
\keyword{perturbage.}
