% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetToGct1.3.R
\name{esetToGct1.3}
\alias{esetToGct1.3}
\title{A function to convert an eset (expression set class) to a "gct" file format}
\usage{
esetToGct1.3(eset, gctFile = NA, path_to_write)
}
\arguments{
\item{eset}{The expression set (expression set class from Biobase).}

\item{gctFile}{This optional parameter specifies the resulting "gct" file name. If not provided, the experiment name will be used.}

\item{path_to_write}{This parameter specifies the path to save the "gct" format of the downloaded data.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to convert an eset (Biobase class object) to a "gct" format (text file) and save the ".gct" file.
}
\examples{
## not run
mygct <- esetToGct1.3(eset=eset, gctFile="my_gct_file", path_to_write="your/path/to/write/here")
## end not run
}
\keyword{gct}
