% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expInfo.R
\name{expInfo}
\alias{expInfo}
\title{A function to retrieve an experiment metadata}
\usage{
expInfo(exp, debuging = FALSE)
}
\arguments{
\item{exp}{The experiment which you want to get metadata for.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to retrieve necessary metadata about an experiment from databases.
}
\examples{
## not run
EDS1014 <- expInfo(exp="EDS-1014")
## end not run
}
\keyword{Download.}
