% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_eset.R
\name{filter_eset}
\alias{filter_eset}
\title{A function for filtering Biobase ExpressionSets.}
\usage{
filter_eset(eset, filterchk = NULL, basement = NULL, includeORexclude,
  new.esetName = "filtered.eset", path_to_write = NULL)
}
\arguments{
\item{eset}{should be an ExpressionSet with the class of Biobase Eset containing pData and exprs data}

\item{filterchk}{filterchk is in the form of a string. It can have multiple property of the pData to filter
based on. Each pair of property:value should be separated by a comma ",". Pairs are saparated
by colon ":". An example ot filterchk is :
filterchk="property1:value1,property1:value2,property2:value1"}

\item{basement}{A two value vector containing "baseline" and "treatment".}

\item{includeORexclude}{This argument basically is designed to filter the pData based on what 
filterchk is or selected property of eset when is set to "1", or the reverse selection of filterchk 
when set to "2". 
It also can have NULL or "n" values when there is no property selected, in this case 
the function will return the original ExpressionSet.}

\item{new.esetName}{This parameter should be the eset name, if user wants to have a specific name.}

\item{path_to_write}{This optional parameter specifies the path which user wants to save the 
filtered expressionSet to be saved in as an .RData file.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to filter an ExpressionSet based on specified properties.
}
\examples{
## not run
data(expl_eset)
res <- filter_eset(eset=expl_eset, filterchk="ER:+,ER:-", includeORexclude=1)
## not run
}
\keyword{filter}
