% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes_hgenes_in_platform.R
\name{find_genes_hgenes_in_platform}
\alias{find_genes_hgenes_in_platform}
\title{A function for ...}
\usage{
find_genes_hgenes_in_platform(exp, glist, org = "Hs", debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
