% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConcordancesBing.R
\name{findConcordancesBing}
\alias{findConcordancesBing}
\title{A function for visualizing, analysing and further investigating on a next-gene experiment.}
\usage{
findConcordancesBing(exp = NULL, path_to_write, libName = NULL,
  userUploadedProfile = NULL, queryGenelist = NULL, write = FALSE,
  homoloGenes = TRUE, cutoff = 0.05, logPcut = 10,
  debuging = FALSE, sigDB = "ilincs_sigs",
  algorithm = "wtCor_all_exp", topSigs = FALSE, metadata = FALSE,
  chunk = 500, ncors = 10, org = "Hs")
}
\arguments{
\item{exp}{This argument is the experiment name, defined as a string corresponding to GenomicsPortals experiment name. The default is set to NULL.}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the 
output files (heatmaps, TreeView files and some other output files) to be saved in.}

\item{libName}{libName defines which precomputed signature library to calculate concordances against.}

\item{userUploadedProfile}{Is required if user wants to use his/her own provided gene Profiles in the analysis. It should be in the form of a string containig the path to Profile text file.
This text file should contain geneIDs in the first column, Coefficients in the second and pValues in the third column (It's also possible to submit the file without pValues).}

\item{queryGenelist}{This is a string parameter consisting of user provided gene IDs separated by commas to be used in querying the dataset or experiment.}

\item{write}{write argument is used (in downstreem function "parse_res") to save the SignatureTable for further uses.
It should be set to TRUE, as it is by default, to save the SignatureTable in the form of txt.}

\item{homoloGenes}{homoloGenes To find human homologus genes if signature is not from human.}

\item{cutoff}{Is the p-value cutoff to filter out concordance results.}

\item{debuging}{for debugging purposes only!.}

\item{sigDB}{Which precomputed signature database to use, before or after 2018.}

\item{algorithm}{Different data source and connectivity methods including c("Cor_all_exp", "Cor_all_plog", "Cor_updn_exp", "Cor_updn_plog", "wtCor_all_exp").}

\item{chunk}{Used in splitting sig data for ENCODE only.}

\item{ncors}{Used in multi thread calculation of ENCODE sig data only.}

\item{org}{Organism used for mapping genes.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to run multiple analysis and is one of the main work flows in GenomicsPortals.
}
\examples{
## not run
# For uploading a signature and finding concordances to a library:
res <- findConcordances(exp="sample.tsv", path_to_write="/patho/to/your/output/directory/", libName="LIB_5",
userUploadedProfile="/patho/to/your/output/directory/")

## not run
}
\keyword{algorithm}
\keyword{exp,}
\keyword{userUploadedProfile,}
