% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalPathwayAnalysis_SPIA.R
\name{functionalPathwayAnalysis_SPIA}
\alias{functionalPathwayAnalysis_SPIA}
\title{A function for SPIA algorithm analysis.}
\usage{
functionalPathwayAnalysis_SPIA(sessionID, gpgene, path_to_write,
  fdr = FALSE, cutoff = 0.01, species = "Hs", foldChange = FALSE)
}
\arguments{
\item{sessionID}{This is the sessionID generated by previous "master_LincsDataAnalysis" or "master_diffGenes" functions run.}

\item{gpgene}{A character string consisting of all significant differentially expressed gene IDs (Entrez IDs) separated by commas.}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the 
output files (heatmaps, TreeView files and some other output files) to be saved in.}

\item{fdr}{False Discovery Rate, if provided set to TRUE.}

\item{cutoff}{Significance level (alpha).}

\item{species}{A character string. One of c("Hs", "Mm", "Rn") for Human, Mouse or Rat pathways respectively.}

\item{foldChange}{A logical value if fold change is provided.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to run SPIA method at least 10 times faster using multithreading.
}
\examples{
res <- functionalPathwayAnalysis_SPIA(....)
}
\keyword{SPIA.}
