% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneid2symbol.R
\name{geneid2symbol}
\alias{geneid2symbol}
\title{A function to convert Gene IDs to Gene Symbols}
\usage{
geneid2symbol(idlist, species = "Hs", description = FALSE,
  verbose = FALSE, debuging = FALSE)
}
\arguments{
\item{idlist}{A character string containing a single or list of Entrez gene IDs separated by commas.}

\item{species}{Three different available species in the Genomics Portal database ("Hs" for Human, "Mm" for Mouse or "Rn" for Rat).}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}
}
\description{
This function allows you to convert official Gene IDs (Entrez) to Gene Symbols.
}
\examples{
geneSymbols <- geneid2symbol(idlist = "57535,11010,283987,2920,2113,10612,9619")
}
\keyword{Entrez}
\keyword{Gene}
\keyword{ID}
