% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{genePlot}
\alias{genePlot}
\title{A function for ...}
\usage{
genePlot(sessionID, geneID, prop, path_to_write = "/mnt/raid/tmp/",
  output = "json", fontSize = 13)
}
\arguments{
\item{sessionID}{sessionID generated by previous step of analysis.}

\item{geneID}{Any single gene Entrez ID to create scatter plot.}

\item{path_to_write}{This parameter specifies the path which user wants to save the results.}

\item{output}{Any of "json" or "widget" formats for the plot.}

\item{debuging}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if debuging=TRUE dev servers will be used.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
## Do not run
sessionID <- "Fri_Oct_5_13_16_50_2018_4925482"
geneID <- "2099"
path_to_write <- "/mnt/raid/tmp/"
plt <- genePlot(sessionID, geneID, prop="ER", path_to_write, output="json", debuging=TRUE)
## End do not run
}
\keyword{filter}
