% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSID.R
\name{generateSID}
\alias{generateSID}
\title{A function for generate session identifier.}
\usage{
generateSID()
}
\arguments{
\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to generate a unique session ID at random.
}
\examples{
generateSID()
}
\keyword{...}
