% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutpoints.R
\name{get_cutpoints}
\alias{get_cutpoints}
\title{A function to return percentiles of different signature library concordances}
\usage{
get_cutpoints(libName = c("LIB_1", "LIB_2", "LIB_3", "LIB_5", "LIB_6",
  "LIB_7", "LIB_8", "LIB_9", "LIB_10", "LIB_11", "LIB_12"),
  level = rev(c("pos05", "pos10", "neg10", "neg05")))
}
\arguments{
\item{libName}{The name of signature library user wants to get percentiles for. The default is all.}

\item{level}{any of 0.5, 1.0, 99.0 and 99.5 percentiles equivalent to positive 0.5 and 1.0 percentiles and negative 0.5 and 1.0 percentiles.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to get the 0.5, 1.0, 99.0 and 99.5 percentiles of different concordance signature libraries.
}
\examples{
## not run
get_cutpoints()
# returns all available percentiles for all library concordances.
## end not run
}
\keyword{Concordance.}
