% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_probe_DB_genelists.R
\name{get_gene_probe_DB_genelists}
\alias{get_gene_probe_DB_genelists}
\title{A function for ...}
\usage{
get_gene_probe_DB_genelists(db, genelistid, platform, debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
