% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotatedExp.R
\name{getAnnotatedExp}
\alias{getAnnotatedExp}
\title{A function to download a nicely annotated whole dataset from ilincs.org portal}
\usage{
getAnnotatedExp(exp, id.na.rm = TRUE, debuging = FALSE)
}
\arguments{
\item{exp}{The experiment which you want to filter and download. It should be present in the portal's database.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to download a dataset as "gct" format with gene annotations.
}
\examples{
## not run
mygct <- getAnnotatedExp(exp="EDS-1014")
## end not run
}
\keyword{Download.}
