% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomologousGenes.R
\name{getHomologousGenes}
\alias{getHomologousGenes}
\title{A function for ...}
\usage{
getHomologousGenes(glist, org = "Hs")
}
\arguments{
\item{glist}{A vector of geneIDs.}

\item{tax}{TaxID for the specious ...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to find Human homolog genes for non-Human specious.
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
