% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeptideMeta.R
\name{getPeptideMeta}
\alias{getPeptideMeta}
\title{A function to retrieve peptide info}
\usage{
getPeptideMeta(peptideList, metacols = "*", debuging = FALSE)
}
\arguments{
\item{peptideList}{A vector of LSM IDs for drugs.}

\item{metacols}{Metadata column names needed.}

\item{debuging}{Only for database test purposes.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to retrieve peptide metadata from uniprot.
}
\examples{
## not run
myLSMs <- getPeptideMeta(c("LSM-10001","LSM-10000","LSM-36603","LSM-1535"), minSimilarity=0.2, debuging=T)
## end not run
}
\keyword{Uniprot.}
