% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProp.R
\name{getProp}
\alias{getProp}
\title{A function for ...}
\usage{
getProp(sessionID, path_to_write = "/mnt/raid/tmp/")
}
\arguments{
\item{sessionID}{sessionID generated by previous step of analysis.}

\item{path_to_write}{This parameter specifies the path which user wants to save the results.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
## Do not run
sessionID <- "Fri_Oct_5_13_16_50_2018_4925482"
path_to_write <- "/mnt/raid/tmp/"
allprop <- getProp(sessionID, path_to_write)
## End do not run
}
\keyword{filter}
