% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServerSettings.R
\name{getServerSettings}
\alias{getServerSettings}
\title{A function for choosing database and coresponding server.}
\usage{
getServerSettings(debuging = FALSE, sigDB = "ilincs_sigs",
  chost = "dev.ilincs.org", dhost = "gimm2.ketl.uc.edu",
  dport = 3306)
}
\arguments{
\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{sigSet}{Which signature version user wants to use in the analysis.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to choose a server which provides databases. It is for testing purposes only before a dataset/database is released.
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
