% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignatureMeta.R
\name{getSignatureMeta}
\alias{getSignatureMeta}
\title{A function to extract signature meta data from ilincs.org portal}
\usage{
getSignatureMeta(signatures,
  prop = "signatureID,datasetID,perturbagenID,compound,lincsPertID,integratedGeneTargets,treatment,concentration,cellLine,time,factor,Level1,Level2,antibodyTarget",
  debuging = FALSE, sigDB = "ilincs_sigs")
}
\arguments{
\item{signatures}{A string containing a single or list of signatures exist in ilincs portal separated by commas.}

\item{prop}{A string containing a single or list of meta data columns from following list separated by commas: 
c(libraryID, signatureID, datasetID, platform, factor, Level1, Level2, cID, compound, concentration, cellLine, 
n_trt_samples, n_ctr_samples, antibodyTarget, peakType, treatment, time, lincsPertID, concordanceTable, stitchID,
perturbagenID, pubChemID, treeviewSigId)
By default the function returns 11 most important property of signatures.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to download/extract selected meta data related to specified signatures from ilincs portal.
}
\examples{
## not run
mysig_property <- getSignatureMeta(signatures="LINCSSH_1,LINCSSH_10,LINCSSH_100,LINCSSH_10000,ENC_112,GDS_152")
## end not run
}
\keyword{Signature.}
