% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimmHeat.R
\name{gimmHeat}
\alias{gimmHeat}
\title{A function for ...}
\usage{
gimmHeat(x, gimmTrunc = NA, Rowv = TRUE, Colv = if (symm) "Rowv" else
  TRUE, distfun = dist, hclustfun = hclust, dendrogram = c("both",
  "row", "column", "none"), symm = FALSE, scale = c("none", "row",
  "column"), na.rm = TRUE, revC = identical(Colv, "Rowv"), add.expr,
  breaks, col = "heat.colors", colsep, rowsep, sepcolor = "white",
  sepwidth = c(0.05, 0.05), cellnote, notecex = 1, notecol = "cyan",
  na.color = par("bg"), trace = c("column", "row", "both", "none"),
  tracecol = "cyan", hline = median(breaks), vline = median(breaks),
  linecol = tracecol, margins = c(5, 5), ColSideColors, RowSideColors,
  cexRow = 0.2 + 1/log10(nr), cexCol = 0.2 + 1/log10(nc),
  labRow = NULL, labCol = NULL, key = TRUE, keysize = 1.5,
  density.info = c("histogram", "density", "none"), denscol = tracecol,
  symkey = min(x < 0, na.rm = TRUE), densadj = 0.25, main = NULL,
  xlab = NULL, ylab = NULL, lmat = NULL, lhei = NULL,
  lwid = NULL, ...)
}
\arguments{
\item{na.rm}{...}

\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
