% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSlocal.R
\name{GRSlocal}
\alias{GRSlocal}
\title{A function for ...}
\usage{
GRSlocal(query.p, reference.p, query.gL = NULL, reference.gL = NULL,
  estimateNullDistr = TRUE, nullDistrQuantiles = c(0.9, 0.95, 0.99),
  nullDistrN = 100)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{na.rm}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
