% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilincsInfo.R
\name{ilincsInfo}
\alias{ilincsInfo}
\title{A function for getting ilincs portal information.}
\usage{
ilincsInfo()
}
\arguments{
\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to see what R version, ilincsR version, compute server and user is being used in the analysis.
}
\examples{
ilincsInfo()
}
\keyword{...}
