% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggLinks.R
\name{keggLinks}
\alias{keggLinks}
\title{A function for creating KEGG hyperlinks.}
\usage{
keggLinks(topGenes, pathwayidlist = NULL, fdr = FALSE, cutoff = 0.05,
  species = "Hs", foldChange = FALSE, path_to_write = NULL,
  sessionID = NULL, debuging = FALSE)
}
\arguments{
\item{topGenes}{A data.frame object containing geneIDs as row names, log2 fold Changes and pValue or adj_pValues for each gene.
This data.frame should at least have two columns named "log2FoldChange" and "pval"/"padj".}

\item{pathwayidlist}{A list of all pathways needed to be used in the analysis.}

\item{fdr}{A logical value for False Discovery Rate. If needs to be used, set to TRUE.}

\item{cutoff}{Significance level (alpha).}

\item{species}{A character string. One of c("Hs", "Mm", "Rn") for Human, Mouse or Rat pathways respectively.}

\item{foldChange}{A logical value if fold change is provided.}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the 
output files (heatmaps, TreeView files and some other output files) to be saved in.}

\item{sessionID}{This is the sessionID generated by previous "master_LincsDataAnalysis" or "master_diffGenes" functions run.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to create many KEGG pathway hyperlincs quikly using several cores, if available.
}
\examples{
res <- KEGGlinks(....)
}
\keyword{KEGGlinks}
