% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmConcordances.R
\name{lsmConcordances}
\alias{lsmConcordances}
\title{A function to retrieve drug similarity}
\usage{
lsmConcordances(lsmList, minSimilarity = 0, debuging = FALSE)
}
\arguments{
\item{lsmList}{A vector of LSM IDs for drugs.}

\item{minSimilarity}{A single numeric value for minimum similarity between drugs, default to "0".}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to retrieve necessary metadata about an experiment from databases.
}
\examples{
## not run
myLSMs <- lsmConcordances(c("LSM-10001","LSM-10000","LSM-36603","LSM-1535"), minSimilarity=0.2, debuging=T)
## end not run
}
\keyword{ID.}
\keyword{LSM}
