% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_diffGenes.R
\name{master_diffGenes}
\alias{master_diffGenes}
\title{A function for finding differentially expressed genes in a next-gene experiment.}
\usage{
master_diffGenes(exp, prop, filterchk = "n", basement = NULL,
  includeORexclude, ifCluster = "both", up = 4000, down = 1000,
  window.size = 50, path_to_write, pORqVal, previousID = NULL,
  reAnalyse = FALSE, glist = NULL, pValue = NULL,
  foldchange = NULL, MLinfo = NULL, homoloGenes = TRUE,
  multigroup = FALSE, sigDB = "ilincs_sigs", chost = "ilincs.org",
  debuging = FALSE)
}
\arguments{
\item{exp}{This argument is the experiment name, defined as a string corresponding to GenomicsPortals experiment name. The default is set to NULL.}

\item{prop}{prop defines which property of the experiment is going to be used for statistical analysis to specify the groups of samples.}

\item{filterchk}{filterchk is in the form of a string. It can have multiple property of the pData to filter
based on. Each pair of property:value should be separated by a comma ",". Pairs are saparated
by colon ":". An example ot filterchk is :
filterchk="property1:value1,property1:value2,property2:value1"}

\item{basement}{A two value vector containing "baseline" and "treatment".}

\item{includeORexclude}{This argument basically is designed to filter the pData based on what 
filterchk is or selected property of eset when is set to "1", or the reverse selection of filterchk 
when set to "2". 
It also can have NULL or "n" values when there is no property selected, in this case 
the function will return the original ExpressionSet.}

\item{ifCluster}{This parameter defines in which way the data should be clustered. The default value in this work flow is set to "both" which 
clusters both genes (rows) and samples (columns) in the data matrix.}

\item{up, down}{Specifies the maximum number of up and down regulated genes.}

\item{window.size}{...}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the 
output files (heatmaps, TreeView files and some other output files) to be saved in.}

\item{pORqVal}{is a string argument and defines if "p" values are used in the analysis or "q" values.}

\item{previousID}{This will be used if user wants to reanalyse a new dataset or experiment using the gene list derived from first round of analysis.}

\item{reAnalyse}{Should be set to TRUE only if same/another experiment or dataset is needed to be analysed based on the results of the first round of analysis.}

\item{pValue}{pValue is a numeric value between 0 and 1 and determines the size of the tests (ttest/ftest or the alpha value) 
in selecting differentially expressed genes.}

\item{foldchange}{This is also a numeric value for selecting differentially expressed genes based on their log2 fold changes between groups.
If set to NULL along with pValue of NULL the function will use top 100 genes based on p-values.}

\item{MLinfo}{This argument is a list object containing previous analysis arguments for the function and should be used only if 
reAnalyse is set to TRUE. After each round of running "master_LincsDataAnalysis" function, all the function arguments are saved
just in case user wants to proceed any reAnalysing with a new dataset.}

\item{multigroup}{A critical parameter for specifying which work flow user wants to choose. The default is set to FALSE meaning the function
will run in the two group mode and will calculates the signatures in specified groups (only two groups are allowed). 
In the other hand, if multigroup is set to TRUE, it will run an F test between all selected groups (selected previously in filterchk).}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to run differntial expression analysis in GenomicsPortals.
}
\examples{
## not run
res_multigroup <- master_diffGenes(exp="EDS-1014", prop="subtype", filterchk="subtype:Basal,subtype:Luminal,subtype:Claudin-low", 
basement=NULL, includeORexclude="1", ifCluster="both", pORqVal="p", pValue=0.000001, foldchange=4, 
up=4000, down=1000, window.size=50, path_to_write="/patho/to/your/output/directory/", multigroup=TRUE)

# For reanalysing:
res2_multigroup <- master_diffGenes(exp="EDS-1014", prop="subtype", filterchk="subtype:Basal,subtype:Luminal,subtype:Claudin-low", 
basement=NULL, includeORexclude="1", ifCluster="both", pORqVal="p", pValue=0.000001, foldchange=4, 
up=4000, down=1000, window.size=50, path_to_write="/patho/to/your/output/directory/", multigroup=TRUE, 
previousID="find previousID from your saved results in the first step", reAnalyse=TRUE, MLinfo=MLinfo)

# previousID may look something like this: "Mon_Oct_26_23_49_56_2015_1084923"
# "Date + a randomely generated number separated by under scores"
## not run
}
\keyword{exp,}
\keyword{prop,}
\keyword{reAnalyse.}
