% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_for_servlet.R
\name{master_for_servlet}
\alias{master_for_servlet}
\title{A function for ...}
\usage{
master_for_servlet(db, exp, prop, filterchk, filters, ifCluster,
  includeORexclude, esetName, ifLR, path_to_write, up, down, window.size,
  debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
