% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_GRS.R
\name{master_GRS}
\alias{master_GRS}
\title{A function for "Generalized Random Set" analysis.}
\usage{
master_GRS(data1, data2, src1, src2, prop1, prop2, specie1, specie2,
  filters1 = NULL, filters2 = NULL, filterchk1 = "n",
  filterchk2 = "n", includeORexclude1, includeORexclude2, statusFile,
  debuging = FALSE)
}
\arguments{
\item{data1}{.}

\item{data2}{.}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to perform the concordance analysis between two genome-scale differential expression profiles 
(Query and Reference profiles). The method produces the statistical significance of the overall cocordance without 
requiring specification of significance cutoffs as well as the list of genes contributing to the overall concordance. 
The Query and Reference differential expression profiles can be constructed by the real-time analysis of the datasets 
in Genomics Portals or can be separately uploaded in the form of the gene list with p-values of differential expression.
}
\examples{
## not run
src1 <- src2 <- "path/to/your/files" ## src2 can be set to any of available experiments on iLincs portal
specie1 <- specie2 <- "human" ## or they could be different
prop1 <- prop2 <- "cell_id" ## this was just a test from previous analysis, they should be grab from appropriate experiment
filterchk2 <- "cell_id:MCF7,cell_id:PC3" 
grs <- master_GRS("sample1.txt","sample2.txt",src1,src2,prop1,prop2,specie1,specie2,includeORexclude1=1,includeORexclude2=1,
		statusFile="testGRS")
## end not run
}
\keyword{...}
