% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_LincsDataAnalysis.R
\name{master_LincsDataAnalysis}
\alias{master_LincsDataAnalysis}
\title{A function for visualizing, analysing and further investigating on a next-gene experiment.}
\usage{
master_LincsDataAnalysis(exp = NULL, prop = NULL, filterchk = "n",
  basement = NULL, includeORexclude = 2, ifCluster = "Genes",
  up = 4000, down = 1000, window.size = 50, path_to_write,
  pORqVal = "p", libName = NULL, userUploadedProfilePath = NULL,
  previousID = NULL, reAnalyse = FALSE, queryGenelist = NULL,
  pValue = NULL, foldchange = NULL, write = TRUE,
  homoloGenes = TRUE, debuging = FALSE, multigroup = FALSE,
  sigDB = "ilincs_sigs", chost = "ilincs.org", org = "Hs")
}
\arguments{
\item{exp}{This argument is the experiment name, defined as a string corresponding to GenomicsPortals experiment name. The default is set to NULL.}

\item{prop}{prop defines which property of the experiment is going to be used for statistical analysis to specify the groups of samples.}

\item{filterchk}{filterchk is in the form of a string. It can have multiple property of the pData to filter
based on. Each pair of property:value should be separated by a comma ",". Pairs are saparated
by colon ":". An example ot filterchk is :
filterchk="property1:value1,property1:value2,property2:value1"}

\item{basement}{basement A two value vector containing "baseline" and "treatment".}

\item{includeORexclude}{This argument basically is designed to filter the pData based on what 
filterchk is or selected property of eset when is set to "1", or the reverse selection of filterchk 
when set to "2". 
It also can have NULL or "n" values when there is no property selected, in this case 
the function will return the original ExpressionSet.}

\item{ifCluster}{This parameter defines in which way the data should be clustered. The default value in this work flow is set to "Genes" which clusters only genes or rows in the data matrix.}

\item{up, down}{Specifies the maximum number of up and down regulated genes.}

\item{window.size}{...}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the 
output files (heatmaps, TreeView files and some other output files) to be saved in.}

\item{pORqVal}{Defines if "p" values are used in the analysis or "q" values.}

\item{libName}{Is a parameter to specify one of five Signature libraries in the GenomicsPortals database to calculate the concordance of generated signatures during the analysis against.}

\item{userUploadedProfilePath}{Is required if user wants to use his/her own provided gene Profiles in the analysis. It should be in the form of a string containig the path to Profile text file.
This text file should contain geneIDs in the first column, Coefficients in the second and pValues in the third column (It's also possible to submit the file without pValues).}

\item{previousID}{This will be used if user wants to reanalyse a new dataset or experiment using the gene list derived from first round of analysis.}

\item{reAnalyse}{Should be set to TRUE only if same/another experiment or dataset is needed to be analysed based on the results of the first round of analysis.}

\item{queryGenelist}{This is a string parameter consisting of user provided gene IDs separated by commas to be used in querying the dataset or experiment.}

\item{write}{write argument is used (in downstreem function "parse_res") to save the SignatureTable for further uses.
It should be set to TRUE, as it is by default, to save the SignatureTable in the form of txt.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to run multiple analysis and is one of the main work flows in GenomicsPortals.
}
\examples{
## not run
# For creating a signature:
res <- master_LincsDataAnalysis(exp="EDS-1013", prop="ER", filterchk="ER:+,ER:-", 
basement=NULL, includeORexclude="1", ifCluster="both", pORqVal="p", 
path_to_write="/patho/to/your/output/directory/", 
userUploadedProfilePath=NULL, up=4000,	down=1000, window.size=50)

# For creating a signature and calculating concordances to a library of signatures:
res <- master_LincsDataAnalysis(exp="EDS-1013", prop="ER", filterchk="ER:+,ER:-", 
basement=NULL, includeORexclude="1", ifCluster="both", pORqVal="p", 
path_to_write="/patho/to/your/output/directory/", libName="LIB_5",
userUploadedProfilePath=NULL, up=4000,	down=1000, window.size=50)

# For reanalysing:
res <- master_LincsDataAnalysis(exp="EDS-1013",prop="ER", filterchk= "ER:+,ER:-", 
basement= NULL, includeORexclude="1", ifCluster="both", pORqVal="p", 
path_to_write="/patho/to/your/output/directory/", libName="LIB_5",
userUploadedProfilePath=NULL, up=4000,	down=1000, window.size=50, 
previousID="find previousID from your saved results in the first step", reAnalyse=TRUE)

# previousID may look something like this: "Mon_Oct_26_23_49_56_2015_1084923"
# "Date + a randomely generated number separated by under scores"
# For Random Set (RS) analysis pipeline:
glist <- "57535,11010,283987,2920,2113,10612,9619,3263,2300,1282,148327,84002,8038,114793,4059,6591,3576,51523,25803,8828,4478,
26577,8537,8601,29940,80380,858,4067,718,10809,221061,7043,388,57222,8710,6288,3169,7494,123099,84959,5268,56938,301,
5806,11094,8492,4015,10644,57057,7033,140876,5270,3624,8061,9734,7431,5292,7128,169792,10333,2921,8876,3908,399,10916,
776,3767,5329,83451,1191,25837,199221,2902,6549,309,8991,222,55243,124220,10643,414318,23171,100131187,168544,22996,57134,
2065,55612,26996,6566,2152,27092,54847,857,119,284119,1284,79917,3628,5087"
resRS <- master_LincsDataAnalysis(exp=NULL, libName="LIB_2", path_to_write= "/patho/to/your/output/directory/", queryGenelist=glist)

# For uploading a signature and finding concordances to a library:
res <- master_LincsDataAnalysis(exp="sample.tsv",prop=NULL, filterchk= NULL, 
basement= NULL, includeORexclude=NULL, ifCluster="both", pORqVal="p", 
path_to_write="/patho/to/your/output/directory/", libName="LIB_5",
userUploadedProfilePath="/patho/to/your/output/directory/", up=4000, down=1000, window.size=50)

## not run
}
\keyword{exp,}
\keyword{prop,}
\keyword{reAnalyse.}
