% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanOverGene.R
\name{meanOverGene}
\alias{meanOverGene}
\title{A function to average expression and p-Values over unique genes.}
\usage{
meanOverGene(rawSignature)
}
\arguments{
\item{rawSignature}{A data frame of three or four columns (gene ID, gene Name, coefficients and possibly p-Values).}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to create a signature with unique gene IDs by averaging over multiplicated ones.
}
\examples{
## not run
mysig <-  meanOverGene(mysig)
## end not run
}
\keyword{...}
