% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mehdi_wtcor.R
\name{mehdi_wtcor}
\alias{mehdi_wtcor}
\title{A function to calculate weighted concordance between a signature and a library.}
\usage{
mehdi_wtcor(data1, data2 = NULL, weight = NULL,
  alternative = "two.sided", pvalue = TRUE, topSigs = FALSE)
}
\arguments{
\item{data1}{A numeric vector as the input signature.}

\item{data2}{A numeric vector or a matrix in the same row orther as in data1.}

\item{weight}{A numeric vector or a matrix in the same row orther as in data1 
and data2 and the same number of columns in data2.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- mehdi_wtcor(data1, data2, weight)
}
\keyword{...}
