% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGenesInExp.R
\name{myGenesInExp}
\alias{myGenesInExp}
\title{A function for finding common genes between a list and an experiment.}
\usage{
myGenesInExp(exp, glist, org = "Hs", debuging = FALSE)
}
\arguments{
\item{exp}{...}

\item{glist}{...}

\item{test}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to know how many of the genes in your list are or have homologs in a spesified experiment.
}
\examples{
## Not run
genelist <- "1002,9447,10001,10096,79575"
res <- myGenesInExp(exp="EDS-1014", glist=genelist, debuging=FALSE)
## End, not run
}
\keyword{test}
