% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCols_equals_three.R
\name{nCols_equals_three}
\alias{nCols_equals_three}
\title{A function for ...}
\usage{
nCols_equals_three(sessionID, sortedValues, layoutMatrix, allColors,
  pValue, prop, nFactorLevels, factorLevels, nColors, nObs, sortedFactors,
  path_to_write)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
