% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org2tax.R
\name{org2tax}
\alias{org2tax}
\title{A function for converting organism name to tax ID.}
\usage{
org2tax(org)
}
\arguments{
\item{org}{Character value for organism.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to convert organism character code to tax ID.
}
\examples{
## Not run
tax <- org2tax("Hs")
## End, not run
}
\keyword{test}
