% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_res.R
\name{parse_res}
\alias{parse_res}
\title{A function to parse the results from previos analysis and turn it to a data table}
\usage{
parse_res(res, path_to_write = NULL, write = TRUE)
}
\arguments{
\item{res}{The input value which is in the form of a data frame generated by previous functions.}

\item{path_to_write}{This optional parameter specifies the path to save the parsed results in the form of ".txt" and ".RData" files.}

\item{write}{write argument is used to save the SignatureTable for further uses.
It should be set to TRUE, as it is by default, to save the SignatureTable otherwise no table will be saved.}

\item{Value}{The output of this function is in the form of a data.frame. First column is geneIDs, second column is geneNames, 
third column is coefficients and the last column is their probability values for each gene from statistical analysis (ttest).}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to parse the results from previos analysis generated by "master_LincsDataAnalysis" or "master_diffGenes" functions.
}
\examples{
## not run
sigTable <- parse_res(res)
## end not run
}
\keyword{SignatureTable.}
