% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perPropertyAvg.R
\name{perPropertyAvg}
\alias{perPropertyAvg}
\title{A function for ...}
\usage{
perPropertyAvg(dataClust, nFactorLevels, factorLevels, geneNames, pal,
  allColors, nColors, sessionID, dataType, path_to_write, Pvalues = NULL,
  PvalueCutoff = 0.05, prop, debuging = FALSE, chost = "ilincs.org")
}
\arguments{
\item{dataClust}{...}

\item{nFactorLevels}{...}

\item{factorLevels}{...}

\item{geneNames}{...}

\item{pal, allColors, nColors}{...}

\item{sessionID}{...}

\item{dataType}{...}

\item{path_to_write}{}

\item{Pvalues}{...}

\item{PvalueCutoff}{...}

\item{prop}{prop defines which property of the experiment is going to be used for statistical analysis to specify the groups of samples.}

\item{path_to_write}{This necessary parameter specifies the path which user wants to save the heatmaps output file.}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
...
res <- perPropertyAvg()
}
\keyword{...}
