% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_with_properties_eset.R
\name{plot_with_properties_eset}
\alias{plot_with_properties_eset}
\title{A function for ...}
\usage{
plot_with_properties_eset(eset, geneprobes = NULL, prop, web = FALSE,
  dataType, filterchk, basement, includeORexclude, ifCluster, ifLR, exp,
  db, path_to_write, dataFormat, up, down, window.size,
  diffGenesSelected = FALSE, Pvalues = NULL, PvalueCutoff = 0.05,
  eset4save = NULL, allPvalues = NULL, allcoeffs = NULL,
  allgeneprobes = NULL, MLinfo = NULL, sigDB = "ilincs_sigs",
  chost = "ilincs.org", debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
...
res <- plot_with_properties_eset(....)
}
\keyword{...}
