% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinTargets.R
\name{proteinTargets}
\alias{proteinTargets}
\title{A function for analysing and visualizing signatures.}
\usage{
proteinTargets(pertList, path_to_write, debuging = FALSE,
  sigDB = "ilincs_sigs")
}
\arguments{
\item{path_to_write}{This parameter specifies the path which user wants to save the results.}

\item{sigDB}{The signatures version available on iLincs portal. It can have values "new" or "old". The default is  "new".}

\item{idList}{A list of signature names as a single string/character argument separated by commas.}

\item{libName}{not used so far!}

\item{queryID}{not used so far!}

\item{NoOfGenesForHeatmap}{The number of top genes per signature based on their significance level user wants to keep in the analysis.}

\item{glist}{A list of Entrez gene IDs as a single string/character argument 
separated by commas user wants to investigate in the selected signatures.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to analyse and visualize precomputed signatures available in iLincs portal.
}
\examples{
## not run
idList <- "LINCSKD_43588,LINCSKD_100,LINCSKD_49086,LINCSCP_101"
glist <- "1514,3553,8835,727,23635,23463,8870,2817,2195,3280,23200,10868,6774,2887,5525,3486,23588,
	3122,10513,3693,9019,5747,10444,22926,25976,9961,1029,9053,26511,593,1891,11031,28969,148022,
	6850,79174,23,5321,9842,64764,2263,7016,10123,10775,6118,622,80025,10058,10153,9128,6182,2553,
	9801,9805,10953,9217,4927,5257,1070,10318,3988,5696,90861,2958,23210,891,2050,84159,1282,1845,
	7750,1017"
mastergroup <- master_GroupLincsAnalysis(idList=idList, path_to_write="your/path/here", glist=glist, sigDB="ilincs_sigs")
# It can be easily noticed that all these signatures are present in the old version of signatures
# but not in the new version.
## end not run
}
\keyword{...}
