% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCgMatrixFromEset.R
\name{saveCgMatrixFromEset}
\alias{saveCgMatrixFromEset}
\title{A function to create a modified gct file}
\usage{
saveCgMatrixFromEset(esetName, path_to_write = "/mnt/raid/tmp/")
}
\arguments{
\item{esetName}{Original data in the form of eset.}

\item{path_to_write}{This parameter specifies the path to save the "gct" format of the downloaded data. 
The default is set to a temp folder: "/mnt/raid/tmp/"}

\item{author}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
## not run
mygct <- saveCgMatrixFromEset("some_eset")
## end not run
}
\keyword{Download.}
