% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleGRS.R
\name{scaleGRS}
\alias{scaleGRS}
\title{A function for ...}
\usage{
scaleGRS(query.p, reference.p, query.gL = NULL, reference.gL = NULL,
  na.rm = TRUE, estimateNullDistr = TRUE, nullDistrQuantiles = c(0.9,
  0.95, 0.99), nullDistrN = 100, tolerateWarnings = TRUE,
  pAdjust.method.query = NULL, pAdjust.method.reference = NULL,
  lambda = 0.005, rescale = T, plotRescaled = F)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
