% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_heatmaps.R
\name{static_heatmaps}
\alias{static_heatmaps}
\title{A function for ...}
\usage{
static_heatmaps(sessionID, dataClust, nFactorLevels, factorLevels,
  geneNames, pal, rsd, allColors, prop, nColors, ifCluster, prop_distance,
  isPropNone, path_to_write, dataFormat, sigDB = "ilincs_sigs",
  chost = "ilincs.org", debuging = FALSE)
}
\arguments{
\item{query.p}{...}

\item{reference.p}{...}

\item{query.gL}{...}

\item{reference.gL}{...}

\item{na.rm}{...}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{...}
