% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol2geneid.R
\name{symbol2geneid}
\alias{symbol2geneid}
\title{A function to convert Gene Symbols to Gene IDs}
\usage{
symbol2geneid(symlist, org = "Hs", debuging = FALSE)
}
\arguments{
\item{symlist}{A character string containing a single or list of gene symbols separated by commas.}

\item{test}{This is for database testing purposes only. The default database server used in all work flows is "eh3" but it can be set to 
alternative database server "gimm2" if test="TRUE" to make sure the uploaded dataset/experiment is working fine.}

\item{exact}{Logical, to find exactly matched gene symbols to database. If FALSE, matching is not case sensitive.}

\item{species}{Three different available species in the Genomics Portal database ("Hs" for Human, "Mm" for Mouse or "Rn" for Rat).}
}
\description{
This function allows you to convert official Gene Symbols to Gene IDs (Entrez gene IDs).
}
\examples{
geneSymbols <- geneid2symbol(idlist = "57535,11010,283987,2920,2113,10612,9619")
}
\keyword{Entrez}
\keyword{Gene}
\keyword{ID}
