% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadedSigProcess.R
\name{uploadedSigProcess}
\alias{uploadedSigProcess}
\title{A function to preProcess an uploaded signature file}
\usage{
uploadedSigProcess(exp = NULL,
  userUploadedProfilePath = "/mnt/raid/tmp/",
  path_to_write = "/mnt/raid/tmp/", org = "Hs", write = TRUE,
  topGenes = NULL, debuging = FALSE)
}
\arguments{
\item{exp}{Your uploaded file name. The file format should be in plain text with tab/comma/semicolon/space delimiters (only one of them in a file though)}

\item{userUploadedProfilePath}{The folder path containing your file.}
}
\description{
This function allows you to translate gene IDs to gene Names and vice versa and remove unavailable genes in iLincs portal.
}
\examples{
processedSig <- uploadedSigProcess(exp = "sample.xls", userUploadedProfilePath="/mnt/raid/tmp/")
}
\keyword{Entrez}
\keyword{Gene}
\keyword{ID}
