% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{A function for ...}
\usage{
volcanoPlot(sigfile, sigType = "notENCODE", xlimit = c(-6, 6),
  ylimit = c(0, 10), DiffExpLowerCutOff = -2, DiffExpUpperCutOff = 2,
  pValCutoff = 0.05, labels = FALSE, cex = 0.6,
  path_to_write = "/mnt/raid/tmp/")
}
\arguments{
\item{sigfile}{...}

\item{sigType}{...}

\item{xlimit}{...}

\item{ylimit}{...}

\item{...}{}

\item{estimateNullDistr}{...}

\item{nullDistrQuantiles}{...}

\item{nullDistrN}{...}

\item{tolerateWarnings}{...}

\item{pAdjust.method.query}{...}

\item{pAdjust.method.reference}{...}

\item{lambda}{...}

\item{rescale}{...}

\item{plotRescaled}{...}

\item{multicoreN}{...}

\item{authors}{M. Fazel-Najafabadi}
}
\description{
This function allows you to ...
}
\examples{
res <- batchTregGRS(....)
}
\keyword{test}
